Added Janky Workaround for Adding GoogleBooks Images
I thought of this at midnight :new_moon:  It's a janky workaround because it does NOT account for if the original image (either user upload or the GB Image) already has the word AMPERSAND in the name. I figured this might be a rare case, but it could still happen.